backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
playwright==1.49.0
httpx==0.27.0
anthropic==0.40.0
supabase==2.10.0
redis==5.0.8
rq==1.16.2
python-dotenv==1.0.1
pydantic==2.9.0
resend==2.3.0
beautifulsoup4==4.12.3
lxml==5.3.0
fake-useragent==1.5.1
aiohttp==3.10.0
